Move task and extension definitions into src/index.ts
